Refactor back end code by modifying routes and adding code to there specific files
