+++ helper/users_helper.js
@@ -1,0 +1,9 @@
+const User = require("../models/User");
+
+const UsersHelper = {
+     deleteUserChallenge: (user_id, hunt_id, success, failure)=>{
+       
+     }
+ }
+
+module.exports = UsersHelper;
