+++ routes/api/hunts.js
@@ -11,45 +11,21 @@ var AWS = require("aws-sdk");
 const keys = require('../../config/keys');
 var fs = require('file-system');
 var Category = require('../../models/Category');
-const { HuntsHelper } = require("../../helper/hunts_helper");
 
 router.get("/test", (req, res) => res.json({msg: "This is the hunts route"}));
 
-//fetch challenges based on categories
-//res will be each category id with all challenges under it
-router.get("/", (req, res) => {
-    Category.find()
-    .sort({date: -1})
-    .then(categories => {
-       let combo = {}
-        for (let i = 0; i < categories.length; i++) {
-            Hunt.find({category: categories[i]}).then(challenges=>{
-                combo[categories[i]._id] = challenges;
-                if(i==categories.length-1){
-                    return res.json(combo)
-                }
-            })
-        }
-    }).catch(error => res.status(404).json({error: "No challenges were found"}))
-});
 
-//fetch user created hunts
-router.get("/:user_id", (req, res) => {
-    Hunt.find({user: req.params.user_id})
-    .sort({date: -1})
-    .then(hunts => {
-        return res.json(hunts);
-    }).catch(error => res.status(404).json({error: "No hunts were found"}))
-});
 
-// move to challenges
-// fetch user's completed challenge stats
-router.get("/stats/:user_id", (req, res) => {
-    Challenge.find({user: req.params.user_id})
-    // .sort({date: -1})
-    .then(completedChallenges=> res.json(completedChallenges))
-        .catch(error => res.status(404).json({error: "No challenges were found" }))
-})
+// //fetch user created hunts
+// router.get("/:user_id", (req, res) => {
+//     Hunt.find({user: req.params.user_id})
+//     .sort({date: -1})
+//     .then(hunts => {
+//         return res.json(hunts);
+//     }).catch(error => res.status(404).json({error: "No hunts were found"}))
+// });
+
+
 
 //fetch hunt detail by Id
 router.get("/:id", (req, res) => {
@@ -58,70 +34,20 @@ router.get("/:id", (req, res) => {
     .catch(error => res.status(404).json({error: "No hunt exists with this id"}))
 })
 
-//add challenge to my_challenge list
-router.post('/add/', passport.authenticate('jwt', { session: false }), (req, res) => {
-    User.updateOne(
-        { _id : req.user.id},
-        { $push: {"my_challenges": req.body.hunt_id},
-    })
-    .then((user) => {
-      return  res.json("Challenge Added")
-    })
-    .catch(error => res.status(404).json({error: 'Challenge not added'}))
-})
-
-//remove challenge from my_challenge play list
-router.delete('/', passport.authenticate('jwt', { session: false }), (req, res) => {
-    HuntsHelper.deleteUserChallenge(
-        req.user.id,
-         req.body.hunt_id,
-         (user) => {
-      return  res.json(user)
-    },error => res.status(404).json(error));
-})
-
-//remove hunt from my_hunts list
-//Change the actual url route
-router.delete('/:hunt_id', passport.authenticate("jwt", {session: false}), (req, res) => {
-    User.updateOne(
-        {_id: req.user.id},
-        { $pull: {"my_hunts": req.params.hunt_id}})
-    .then(user => {
-        return res.json(user)
-    })
-    .catch(error => res.status(404).json({error: 'Hunt not removed'}))
-})
-
-//remove user challenges
-router.delete('/my/challenges/:challenge_id', passport.authenticate("jwt", {session: false}), (req, res) => {
-    User.updateOne(
-        {_id: req.user.id},
-        { $pull: {"my_challenges": req.params.challenge_id}})
-    .then(user => {
-        return res.json(user)
-    })
-    .catch(error => res.status(404).json({error: 'Challenge not removed'}))
-})
 
 
-//fetch added challenges details
-router.get("/my/challenges", passport.authenticate('jwt', { session: false }), (req, res) => {
-    User.findById(req.user.id)
-        .sort({ date: -1 })
-        .then(user => {
-            let combo = {}
-            for (let i = 0; i < user.my_challenges.length; i++) {
-                const challengeId = user.my_challenges[i];
-                Hunt.find({ _id: challengeId }).then(challenge => {
-                    combo[challengeId] = challenge[0];
-                    if (i == user.my_challenges.length - 1) {
-                        return res.json(combo)
-                    }
-                })
-            }
-        }).catch(error => res.status(404).json({ error: "No challenges were found" }))
-});
 
+
+// //remove user challenges
+// router.delete('/my/challenges/:challenge_id', passport.authenticate("jwt", {session: false}), (req, res) => {
+//     User.updateOne(
+//         {_id: req.user.id},
+//         { $pull: {"my_challenges": req.params.challenge_id}})
+//     .then(user => {
+//         return res.json(user)
+//     })
+//     .catch(error => res.status(404).json({error: 'Challenge not removed'}))
+// })
 
 router.post("/", [passport.authenticate('jwt', {session: false}), upload.array('photo_collection', 10)], (req, res) => {
     const {errors, isValid} = validateHuntInput(req);
